+++ src/index.ts
@@ -1,4 +1,4 @@
-export const BPD_TOOLKIT_VERSION = "1.2.0";
+export const BPD_TOOLKIT_VERSION = "1.3.0";
 /**
  * Checks if value is undefined
  * @param val value