+++ tests/validator.tests.ts
@@ -1,4 +1,5 @@
 import {
+	compare,
 	equal,
 	//	exists,
 	field,
@@ -10,6 +11,7 @@ import {
 	schema,
 	validate,
 	validateSingleValue,
+	ValidationCallback,
 } from "../src/validator";
 // describe("Tests checking validator", () => {
 
@@ -17,71 +19,71 @@ import {
 
 describe("Tests checking validation method [min]", () => {
 	it("Checks number larger than min", () => {
-		const minFn = min(3);
-		const res = minFn.callback(4);
+		const minFn = min<any>(3);
+		const res = minFn.callback(4, null, null);
 
 		expect(res).toBeTrue();
 	});
 
 	it("Checks number smaller than min", () => {
-		const minFn = min(3);
-		const res = minFn.callback(2);
+		const minFn = min<any>(3);
+		const res = minFn.callback(2, null, null);
 
 		expect(res).toBeFalse();
 	});
 
 	it("Checks number equal to min", () => {
-		const minFn = min(3);
-		const res = minFn.callback(3);
+		const minFn = min<any>(3);
+		const res = minFn.callback(3, null, null);
 
 		expect(res).toBeTrue();
 	});
 
 	it("Checks string length larger than min", () => {
-		const minFn = min(3);
-		const res = minFn.callback("xxxx");
+		const minFn = min<any>(3);
+		const res = minFn.callback("xxxx", null, null);
 
 		expect(res).toBeTrue();
 	});
 
 	it("Checks string length smaller than min", () => {
-		const minFn = min(3);
-		const res = minFn.callback("xx");
+		const minFn = min<any>(3);
+		const res = minFn.callback("xx", null, null);
 
 		expect(res).toBeFalse();
 	});
 
 	it("Checks string equal to min", () => {
-		const minFn = min(3);
-		const res = minFn.callback("xxx");
+		const minFn = min<any>(3);
+		const res = minFn.callback("xxx", null, null);
 
 		expect(res).toBeTrue();
 	});
 
 	it("Checks array length larger than min", () => {
-		const minFn = min(3);
-		const res = minFn.callback([1, 2, 3, 4]);
+		const minFn = min<any>(3);
+		const res = minFn.callback([1, 2, 3, 4], null, null);
 
 		expect(res).toBeTrue();
 	});
 
 	it("Checks array length smaller than min", () => {
-		const minFn = min(3);
-		const res = minFn.callback([1, 2]);
+		const minFn = min<any>(3);
+		const res = minFn.callback([1, 2], null, null);
 
 		expect(res).toBeFalse();
 	});
 
 	it("Checks array equal to min", () => {
-		const minFn = min(3);
-		const res = minFn.callback([1, 2, 3]);
+		const minFn = min<any>(3);
+		const res = minFn.callback([1, 2, 3], null, null);
 
 		expect(res).toBeTrue();
 	});
 
 	it("Checks boolean length smaller than min", () => {
-		const minFn = min(3);
-		const res = minFn.callback(true);
+		const minFn = min<any>(3);
+		const res = minFn.callback(true, null, null);
 
 		expect(res).toBeFalse();
 	});
@@ -89,64 +91,64 @@ describe("Tests checking validation method [min]", () => {
 
 describe("Tests checking validation method [max]", () => {
 	it("Checks number smaller than max", () => {
-		const minFn = max(5);
-		const res = minFn.callback(4);
+		const minFn = max<any>(5);
+		const res = minFn.callback(4, null, null);
 
 		expect(res).toBeTrue();
 	});
 
 	it("Checks number larger than max", () => {
-		const minFn = max(5);
-		const res = minFn.callback(6);
+		const minFn = max<any>(5);
+		const res = minFn.callback(6, null, null);
 
 		expect(res).toBeFalse();
 	});
 
 	it("Checks number equal to max", () => {
-		const minFn = max(5);
-		const res = minFn.callback(5);
+		const minFn = max<any>(5);
+		const res = minFn.callback(5, null, null);
 
 		expect(res).toBeTrue();
 	});
 
 	it("Checks string smaller than max", () => {
-		const minFn = max(5);
-		const res = minFn.callback("xxxx");
+		const minFn = max<any>(5);
+		const res = minFn.callback("xxxx", null, null);
 
 		expect(res).toBeTrue();
 	});
 
 	it("Checks string larger than max", () => {
-		const minFn = max(5);
-		const res = minFn.callback("xxxxxxxx");
+		const minFn = max<any>(5);
+		const res = minFn.callback("xxxxxxxx", null, null);
 
 		expect(res).toBeFalse();
 	});
 
 	it("Checks string equal to max", () => {
-		const minFn = max(5);
-		const res = minFn.callback("xxxxx");
+		const minFn = max<any>(5);
+		const res = minFn.callback("xxxxx", null, null);
 
 		expect(res).toBeTrue();
 	});
 
 	it("Checks array smaller than max", () => {
-		const minFn = max(5);
-		const res = minFn.callback([1, 2, 3, 4]);
+		const minFn = max<any>(5);
+		const res = minFn.callback([1, 2, 3, 4], null, null);
 
 		expect(res).toBeTrue();
 	});
 
 	it("Checks array larger than max", () => {
-		const minFn = max(5);
-		const res = minFn.callback([1, 2, 3, 4, 5, 6, 7]);
+		const minFn = max<any>(5);
+		const res = minFn.callback([1, 2, 3, 4, 5, 6, 7], null, null);
 
 		expect(res).toBeFalse();
 	});
 
 	it("Checks array equal to max", () => {
-		const minFn = max(5);
-		const res = minFn.callback([1, 2, 3, 4, 5]);
+		const minFn = max<any>(5);
+		const res = minFn.callback([1, 2, 3, 4, 5], null, null);
 
 		expect(res).toBeTrue();
 	});
@@ -155,36 +157,36 @@ describe("Tests checking validation method [max]", () => {
 describe("Tests checking validation method [range]", () => {
 	// Numbers
 	it("Checks number smaller than range", () => {
-		const minFn = range(2, 5);
-		const res = minFn.callback(1);
+		const minFn = range<any>(2, 5);
+		const res = minFn.callback(1, null, null);
 
 		expect(res).toBeFalse();
 	});
 
 	it("Checks number larger than range", () => {
-		const minFn = range(2, 5);
-		const res = minFn.callback(6);
+		const minFn = range<any>(2, 5);
+		const res = minFn.callback(6, null, null);
 
 		expect(res).toBeFalse();
 	});
 
 	it("Checks number equal to min range", () => {
-		const minFn = range(2, 5);
-		const res = minFn.callback(2);
+		const minFn = range<any>(2, 5);
+		const res = minFn.callback(2, null, null);
 
 		expect(res).toBeTrue();
 	});
 
 	it("Checks number equal to max range", () => {
-		const minFn = range(2, 5);
-		const res = minFn.callback(5);
+		const minFn = range<any>(2, 5);
+		const res = minFn.callback(5, null, null);
 
 		expect(res).toBeTrue();
 	});
 
 	it("Checks number within range", () => {
-		const minFn = range(2, 5);
-		const res = minFn.callback(4);
+		const minFn = range<any>(2, 5);
+		const res = minFn.callback(4, null, null);
 
 		expect(res).toBeTrue();
 	});
@@ -192,36 +194,36 @@ describe("Tests checking validation method [range]", () => {
 	// String
 
 	it("Checks string smaller than range", () => {
-		const minFn = range(2, 5);
-		const res = minFn.callback("x");
+		const minFn = range<any>(2, 5);
+		const res = minFn.callback("x", null, null);
 
 		expect(res).toBeFalse();
 	});
 
 	it("Checks string larger than range", () => {
-		const minFn = range(2, 5);
-		const res = minFn.callback("xxxxxx");
+		const minFn = range<any>(2, 5);
+		const res = minFn.callback("xxxxxx", null, null);
 
 		expect(res).toBeFalse();
 	});
 
 	it("Checks string equal to min range", () => {
-		const minFn = range(2, 5);
-		const res = minFn.callback("xx");
+		const minFn = range<any>(2, 5);
+		const res = minFn.callback("xx", null, null);
 
 		expect(res).toBeTrue();
 	});
 
 	it("Checks string equal to max range", () => {
-		const minFn = range(2, 5);
-		const res = minFn.callback("xxxxx");
+		const minFn = range<any>(2, 5);
+		const res = minFn.callback("xxxxx", null, null);
 
 		expect(res).toBeTrue();
 	});
 
 	it("Checks string within range", () => {
-		const minFn = range(2, 5);
-		const res = minFn.callback("xxxx");
+		const minFn = range<any>(2, 5);
+		const res = minFn.callback("xxxx", null, null);
 
 		expect(res).toBeTrue();
 	});
@@ -229,36 +231,36 @@ describe("Tests checking validation method [range]", () => {
 	// Arrays
 
 	it("Checks array smaller than range", () => {
-		const minFn = range(2, 5);
-		const res = minFn.callback([1]);
+		const minFn = range<any>(2, 5);
+		const res = minFn.callback([1], null, null);
 
 		expect(res).toBeFalse();
 	});
 
 	it("Checks array larger than range", () => {
-		const minFn = range(2, 5);
-		const res = minFn.callback([1, 2, 3, 4, 5, 6]);
+		const minFn = range<any>(2, 5);
+		const res = minFn.callback([1, 2, 3, 4, 5, 6], null, null);
 
 		expect(res).toBeFalse();
 	});
 
 	it("Checks array equal to min range", () => {
-		const minFn = range(2, 5);
-		const res = minFn.callback([1, 2]);
+		const minFn = range<any>(2, 5);
+		const res = minFn.callback([1, 2], null, null);
 
 		expect(res).toBeTrue();
 	});
 
 	it("Checks array equal to max range", () => {
-		const minFn = range(2, 5);
-		const res = minFn.callback([1, 2, 3, 4, 5]);
+		const minFn = range<any>(2, 5);
+		const res = minFn.callback([1, 2, 3, 4, 5], null, null);
 
 		expect(res).toBeTrue();
 	});
 
 	it("Checks array within range", () => {
-		const minFn = range(2, 5);
-		const res = minFn.callback([1, 2, 3, 4]);
+		const minFn = range<any>(2, 5);
+		const res = minFn.callback([1, 2, 3, 4], null, null);
 
 		expect(res).toBeTrue();
 	});
@@ -266,115 +268,144 @@ describe("Tests checking validation method [range]", () => {
 
 describe("Tests checking validation method [equal]", () => {
 	it("Checks two equal numbers", () => {
-		const fnObj = equal(2);
-		const result = fnObj.callback(2);
+		const fnObj = equal<any>(2);
+		const result = fnObj.callback(2, null, null);
 		expect(result).toBeTrue();
 	});
 
 	it("Checks two different numbers", () => {
-		const fnObj = equal(2);
-		const result = fnObj.callback(3);
+		const fnObj = equal<any>(2);
+		const result = fnObj.callback(3, null, null);
 		expect(result).toBeFalse();
 	});
 
 	it("Checks two equal strings", () => {
-		const fnObj = equal("x");
-		const result = fnObj.callback("x");
+		const fnObj = equal<any>("x");
+		const result = fnObj.callback("x", null, null);
 		expect(result).toBeTrue();
 	});
 
 	it("Checks two different strings", () => {
-		const fnObj = equal("x");
-		const result = fnObj.callback("xx");
+		const fnObj = equal<any>("x");
+		const result = fnObj.callback("xx", null, null);
 		expect(result).toBeFalse();
 	});
 
 	it("Checks two equal booleans", () => {
-		const fnObj = equal(true);
-		const result = fnObj.callback(true);
+		const fnObj = equal<any>(true);
+		const result = fnObj.callback(true, null, null);
 		expect(result).toBeTrue();
 	});
 
 	it("Checks two different booleans", () => {
-		const fnObj = equal(true);
-		const result = fnObj.callback(false);
+		const fnObj = equal<any>(true);
+		const result = fnObj.callback(false, null, null);
 		expect(result).toBeFalse();
 	});
 
 	it("Checks two different arrays", () => {
-		const fnObj = equal([1, 2]);
-		const result = fnObj.callback([1, 2, 3]);
+		const fnObj = equal<any>([1, 2]);
+		const result = fnObj.callback([1, 2, 3], null, null);
 		expect(result).toBeFalse();
 	});
 });
 
 describe("Tests checking validation method [ofType]", () => {
 	it("Returns true when type matches", () => {
-		const fnObj = ofType("string");
-		const result = fnObj.callback("XX");
+		const fnObj = ofType<any>("string");
+		const result = fnObj.callback("XX", null, null);
 		expect(result).toBeTrue();
 	});
 
 	it("Returns false when type does not match", () => {
-		const fnObj = ofType("boolean");
-		const result = fnObj.callback("XX");
+		const fnObj = ofType<any>("boolean");
+		const result = fnObj.callback("XX", null, null);
 		expect(result).toBeFalse();
 	});
 });
 
 describe("Tests checking validation method [match]", () => {
 	it("Matches string to string", () => {
-		const fnObj = match("xx");
-		const result = fnObj.callback("xx");
+		const fnObj = match<any>("xx");
+		const result = fnObj.callback("xx", null, null);
 
 		expect(result).toBeTrue();
 	});
 
 	it("Does not match string to string", () => {
-		const fnObj = match("xxxx");
-		const result = fnObj.callback("xx");
+		const fnObj = match<any>("xxxx");
+		const result = fnObj.callback("xx", null, null);
 
 		expect(result).toBeFalse();
 	});
 
 	it("Matches boolean to string", () => {
-		const fnObj = match("true");
-		const result = fnObj.callback(true);
+		const fnObj = match<any>("true");
+		const result = fnObj.callback(true, null, null);
 
 		expect(result).toBeTrue();
 	});
 
 	it("Does not match boolean to string", () => {
-		const fnObj = match("true");
-		const result = fnObj.callback(false);
+		const fnObj = match<any>("true");
+		const result = fnObj.callback(false, null, null);
 
 		expect(result).toBeFalse();
 	});
 
 	it("Matches Regex to string", () => {
-		const fnObj = match(/x+y$/);
-		const result = fnObj.callback("xxy");
+		const fnObj = match<any>(/x+y$/);
+		const result = fnObj.callback("xxy", null, null);
 
 		expect(result).toBeTrue();
 	});
 
 	it("Does not match Regex to string", () => {
-		const fnObj = match(/x+y$/);
-		const result = fnObj.callback("xx");
+		const fnObj = match<any>(/x+y$/);
+		const result = fnObj.callback("xx", null, null);
 
 		expect(result).toBeFalse();
 	});
 
 	it("Matches Regex to boolean", () => {
-		const fnObj = match(/t..e$/);
-		const result = fnObj.callback(true);
+		const fnObj = match<any>(/t..e$/);
+		const result = fnObj.callback(true, null, null);
 
 		expect(result).toBeTrue();
 	});
 
 	it("Does not match Regex to boolean", () => {
-		const fnObj = match(/t..e$/);
-		const result = fnObj.callback(false);
+		const fnObj = match<any>(/t..e$/);
+		const result = fnObj.callback(false, null, null);
+
+		expect(result).toBeFalse();
+	});
+});
+
+describe("Tests checking method [compare]", () => {
+	const testObject = {
+		name: "name",
+		name2: "name",
+		lastname: "lastname",
+	};
+
+	it("Compares two fields with the same value", () => {
+		const fnObj = compare<any>("name2");
+		const result = fnObj.callback("name", "name", testObject);
+
+		expect(result).toBeTrue();
+	});
+
+	it("Compares two fields with different values", () => {
+		const fnObj = compare<any>("lastname");
+		const result = fnObj.callback("name", "name", testObject);
+
+		expect(result).toBeFalse();
+	});
+
+	it("Compares two fields with different values where comparable does not exist", () => {
+		const fnObj = compare<any>("xxx");
+		const result = fnObj.callback("name", "name", testObject);
 
 		expect(result).toBeFalse();
 	});
@@ -382,15 +413,28 @@ describe("Tests checking validation method [match]", () => {
 
 describe("Tests checking function [validateSingleValue]", () => {
 	it("Returns true when all callbacks pass", () => {
-		const validationCallbacks = [min(2), match("xx")];
-		const result = validateSingleValue("prop", "xxxx", validationCallbacks);
+		const validationCallbacks: ValidationCallback<any>[] = [
+			min(2),
+			match("xx"),
+		];
+		const result = validateSingleValue<any>(
+			"prop",
+			"xxxx",
+			null,
+			validationCallbacks
+		);
 
 		expect(result.result).toBeTrue();
 	});
 
 	it("Returns false when at least one callback fails", () => {
 		const validationCallbacks = [min(2), match("xxy")];
-		const result = validateSingleValue("prop", "xxxx", validationCallbacks);
+		const result = validateSingleValue<any>(
+			"prop",
+			"xxxx",
+			null,
+			validationCallbacks
+		);
 
 		expect(result.result).toBeFalse();
 		expect(result.error).toBeDefined();
@@ -399,7 +443,12 @@ describe("Tests checking function [validateSingleValue]", () => {
 
 	it("Breaks at first error", () => {
 		const validationCallbacks = [ofType("number"), min(2), match("xxy")];
-		const result = validateSingleValue("prop", "xxxx", validationCallbacks);
+		const result = validateSingleValue<any>(
+			"prop",
+			"xxxx",
+			null,
+			validationCallbacks
+		);
 
 		expect(result.result).toBeFalse();
 		expect(result.error).toBeDefined();
@@ -408,9 +457,10 @@ describe("Tests checking function [validateSingleValue]", () => {
 
 	it("Performs all checks when option is set", () => {
 		const validationCallbacks = [ofType("number"), min(2), match("xxy")];
-		const result = validateSingleValue(
+		const result = validateSingleValue<any>(
 			"prop",
 			"xxxx",
+			null,
 			validationCallbacks,
 			{
 				checkAll: true,